Delete ImageSubregionAccessor.cpp
This system is not efficient.
first, each sub-element shouldnt have their own pointer, the master image should be the only element with a pointer, and a unique_ptr at that. Thus, each sub-element should have an offset into the master image's pointer instead.
With that taken into account the subscript operator accessing scheme, it would require 7 function calls to execute a procedure instead of 3 or even 2. 3 to access the sub-element's offset via the subscript operator, 3 more to access the data size, and another for actually calling the function in which you needed the previous information for.
A more efficient approach would instead be to have one array of struct{offset, data size} and access the specific element via manually extracting the specificying the element in the array; F{mip*totalTextureArrayElements*totalFaceElements + textureArrayElem*totalFaceElements + face - 1}.
E.g. mip0/tArrayElem1/face5 = 0*2*6 + 1*6 + 5-1 = 10
       mip1/tArrayElem0/face6 = 1*2*6 + 0*6 + 6-1 = 17